backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bson==0.5.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ctransformers==0.2.27
dataclasses-json==0.6.4
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.41
MarkupSafe==2.1.5
marshmallow==3.21.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
orjson==3.10.0
packaging==23.2
pillow==10.3.0
pinecone-client==3.2.2
pinecone-text==0.9.0
py-cpuinfo==9.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.6.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
structlog==24.1.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wget==3.2
yarl==1.9.4
